% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.waenet.R
\name{cv.waenet}
\alias{cv.waenet}
\title{TODO}
\usage{
cv.waenet(x, y, pf, adWeight, weight, nlambda = 100,
  lambda.min.ratio = 0.001, lambda = NULL, nfolds = 10,
  foldid = NULL, maxit = 1000, eps = 1e-05)
}
\arguments{
\item{x}{A list of \code{m} \code{n x p} numeric matrices. No matrix should
contain an intercept, or any missing values}

\item{y}{A list of \code{m} length n numeric response vectors. No vector
should contain missing values}

\item{pf}{Penalty factor. TODO}

\item{adWeight}{TODO}

\item{nlambda}{Length of automatically generated 'lambda' sequence. If
lambda' is non NULL, 'nlambda' is ignored. Default is 100}

\item{lambda.min.ratio}{Ratio that determines the minimum value of 'lambda'
when automatically generating a 'lambda' sequence. If 'lambda' is not
NULL, 'lambda.min.ratio' is ignored. Default is 1e-3}

\item{lambda}{Optional numeric vector of lambdas to fit. If NULL,
\code{galasso} will automatically generate a lambda sequence based off
of \code{nlambda} and code{lambda.min.ratio}. Default is NULL}

\item{nfolds}{Number of folds to use for cross validation. Default is 10}

\item{maxit}{Maximum number of iterations to run. Default is 1000}

\item{eps}{Tolerance for convergence. Default is 1e-5}
}
\value{
TODO
}
\description{
TODO
}
