% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mianet.df}
\alias{mianet.df}
\title{Synthetic Example Data For mianet}
\format{A data.frame with 500 observations on 21 variables:
\describe{
  \item{Y}{Binary response.}
  \item{X1-X20}{Covariates with missing data.}
}}
\usage{
mianet.df
}
\description{
This synthetic data is taken from the first simulation case from the mianet
paper
}
\keyword{datasets}
